Trasferiti metodi del professionByPersonName
al ProfessionService e controller
